@@ -12,7 +12,7 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
-	"sort"
+	"slices"
 	"strconv"
 	"strings"
 	"testing"
@@ -72,7 +72,7 @@ func TestDirent(t *testing.T) {
 		}
 	}
 
-	sort.Strings(names)
+	slices.Sort(names)
 	t.Logf("names: %q", names)
 
 	if len(names) != 10 {
@@ -107,7 +107,7 @@ func TestDirentRepeat(t *testing.T) {
 	d := t.TempDir()
 
 	var files []string
-	for i := 0; i < N; i++ {
+	for i := range N {
 		files = append(files, fmt.Sprintf("file%d", i))
 	}
 	for _, file := range files {
@@ -145,9 +145,9 @@ func TestDirentRepeat(t *testing.T) {
 	}
 
 	// Check results
-	sort.Strings(files)
-	sort.Strings(files2)
-	if strings.Join(files, "|") != strings.Join(files2, "|") {
+	slices.Sort(files)
+	slices.Sort(files2)
+	if !slices.Equal(files, files2) {
 		t.Errorf("bad file list: want\n%q\ngot\n%q", files, files2)
 	}
 }
