@@ -15,21 +15,21 @@ import (
 func TestFdSet(t *testing.T) {
 	var fdSet unix.FdSet
 	fdSet.Zero()
-	for fd := 0; fd < unix.FD_SETSIZE; fd++ {
+	for fd := range unix.FD_SETSIZE {
 		if fdSet.IsSet(fd) {
 			t.Fatalf("Zero did not clear fd %d", fd)
 		}
 		fdSet.Set(fd)
 	}
 
-	for fd := 0; fd < unix.FD_SETSIZE; fd++ {
+	for fd := range unix.FD_SETSIZE {
 		if !fdSet.IsSet(fd) {
 			t.Fatalf("IsSet(%d): expected true, got false", fd)
 		}
 	}
 
 	fdSet.Zero()
-	for fd := 0; fd < unix.FD_SETSIZE; fd++ {
+	for fd := range unix.FD_SETSIZE {
 		if fdSet.IsSet(fd) {
 			t.Fatalf("Zero did not clear fd %d", fd)
 		}
@@ -39,7 +39,7 @@ func TestFdSet(t *testing.T) {
 		fdSet.Set(fd)
 	}
 
-	for fd := 0; fd < unix.FD_SETSIZE; fd++ {
+	for fd := range unix.FD_SETSIZE {
 		if fd&0x1 == 0x1 {
 			if !fdSet.IsSet(fd) {
 				t.Fatalf("IsSet(%d): expected true, got false", fd)
@@ -55,7 +55,7 @@ func TestFdSet(t *testing.T) {
 		fdSet.Clear(fd)
 	}
 
-	for fd := 0; fd < unix.FD_SETSIZE; fd++ {
+	for fd := range unix.FD_SETSIZE {
 		if fdSet.IsSet(fd) {
 			t.Fatalf("Clear(%d) did not clear fd", fd)
 		}
