@@ -10,8 +10,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"sort"
-	"strings"
+	"slices"
 	"testing"
 
 	"golang.org/x/sys/unix"
@@ -71,9 +70,9 @@ func testGetdirentries(t *testing.T, count int) {
 		}
 	}
 
-	sort.Strings(names)
-	sort.Strings(names2)
-	if strings.Join(names, ":") != strings.Join(names2, ":") {
+	slices.Sort(names)
+	slices.Sort(names2)
+	if !slices.Equal(names, names2) {
 		t.Errorf("names don't match\n names: %q\nnames2: %q", names, names2)
 	}
 }
