@@ -121,6 +121,7 @@ struct termios2 {
 #include <linux/hidraw.h>
 #include <linux/icmp.h>
 #include <linux/icmpv6.h>
+#include <linux/if_addrlabel.h>
 #include <linux/if_alg.h>
 #include <linux/if_bridge.h>
 #include <linux/if_packet.h>
@@ -134,6 +135,7 @@ struct termios2 {
 #include <linux/landlock.h>
 #include <linux/loop.h>
 #include <linux/lwtunnel.h>
+#include <linux/mempolicy.h>
 #include <linux/mpls_iptunnel.h>
 #include <linux/ncsi.h>
 #include <linux/net_namespace.h>
@@ -878,6 +880,8 @@ const (
 	IFA_FLAGS          = C.IFA_FLAGS
 	IFA_RT_PRIORITY    = C.IFA_RT_PRIORITY
 	IFA_TARGET_NETNSID = C.IFA_TARGET_NETNSID
+	IFAL_LABEL         = C.IFAL_LABEL
+	IFAL_ADDRESS       = C.IFAL_ADDRESS
 	RT_SCOPE_UNIVERSE  = C.RT_SCOPE_UNIVERSE
 	RT_SCOPE_SITE      = C.RT_SCOPE_SITE
 	RT_SCOPE_LINK      = C.RT_SCOPE_LINK
@@ -935,6 +939,7 @@ const (
 	SizeofRtAttr       = C.sizeof_struct_rtattr
 	SizeofIfInfomsg    = C.sizeof_struct_ifinfomsg
 	SizeofIfAddrmsg    = C.sizeof_struct_ifaddrmsg
+	SizeofIfAddrlblmsg = C.sizeof_struct_ifaddrlblmsg
 	SizeofIfaCacheinfo = C.sizeof_struct_ifa_cacheinfo
 	SizeofRtMsg        = C.sizeof_struct_rtmsg
 	SizeofRtNexthop    = C.sizeof_struct_rtnexthop
@@ -955,6 +960,8 @@ type RtAttr C.struct_rtattr
 type IfInfomsg C.struct_ifinfomsg
 
 type IfAddrmsg C.struct_ifaddrmsg
+
+type IfAddrlblmsg C.struct_ifaddrlblmsg
 
 type IfaCacheinfo C.struct_ifa_cacheinfo
 
@@ -2293,8 +2300,11 @@ const (
 	NFT_PAYLOAD_LL_HEADER             = C.NFT_PAYLOAD_LL_HEADER
 	NFT_PAYLOAD_NETWORK_HEADER        = C.NFT_PAYLOAD_NETWORK_HEADER
 	NFT_PAYLOAD_TRANSPORT_HEADER      = C.NFT_PAYLOAD_TRANSPORT_HEADER
+	NFT_PAYLOAD_INNER_HEADER          = C.NFT_PAYLOAD_INNER_HEADER
+	NFT_PAYLOAD_TUN_HEADER            = C.NFT_PAYLOAD_TUN_HEADER
 	NFT_PAYLOAD_CSUM_NONE             = C.NFT_PAYLOAD_CSUM_NONE
 	NFT_PAYLOAD_CSUM_INET             = C.NFT_PAYLOAD_CSUM_INET
+	NFT_PAYLOAD_CSUM_SCTP             = C.NFT_PAYLOAD_CSUM_SCTP
 	NFT_PAYLOAD_L4CSUM_PSEUDOHDR      = C.NFT_PAYLOAD_L4CSUM_PSEUDOHDR
 	NFTA_PAYLOAD_UNSPEC               = C.NFTA_PAYLOAD_UNSPEC
 	NFTA_PAYLOAD_DREG                 = C.NFTA_PAYLOAD_DREG
@@ -2381,6 +2391,11 @@ const (
 	NFT_CT_AVGPKT                     = C.NFT_CT_AVGPKT
 	NFT_CT_ZONE                       = C.NFT_CT_ZONE
 	NFT_CT_EVENTMASK                  = C.NFT_CT_EVENTMASK
+	NFT_CT_SRC_IP                     = C.NFT_CT_SRC_IP
+	NFT_CT_DST_IP                     = C.NFT_CT_DST_IP
+	NFT_CT_SRC_IP6                    = C.NFT_CT_SRC_IP6
+	NFT_CT_DST_IP6                    = C.NFT_CT_DST_IP6
+	NFT_CT_ID                         = C.NFT_CT_ID
 	NFTA_CT_UNSPEC                    = C.NFTA_CT_UNSPEC
 	NFTA_CT_DREG                      = C.NFTA_CT_DREG
 	NFTA_CT_KEY                       = C.NFTA_CT_KEY
@@ -3105,8 +3120,25 @@ const (
 )
 
 // generated by:
-// perl -nlE '/^\s*(RTNLGRP_\w+)/ && say "$1 = C.$1"' include/uapi/linux/rtnetlink.h
+// perl -nlE '/^\s*((RTNLGRP_|TCA_)\w+)/ && say "$1 = C.$1"' include/uapi/linux/rtnetlink.h
 const (
+	TCA_UNSPEC            = C.TCA_UNSPEC
+	TCA_KIND              = C.TCA_KIND
+	TCA_OPTIONS           = C.TCA_OPTIONS
+	TCA_STATS             = C.TCA_STATS
+	TCA_XSTATS            = C.TCA_XSTATS
+	TCA_RATE              = C.TCA_RATE
+	TCA_FCNT              = C.TCA_FCNT
+	TCA_STATS2            = C.TCA_STATS2
+	TCA_STAB              = C.TCA_STAB
+	TCA_PAD               = C.TCA_PAD
+	TCA_DUMP_INVISIBLE    = C.TCA_DUMP_INVISIBLE
+	TCA_CHAIN             = C.TCA_CHAIN
+	TCA_HW_OFFLOAD        = C.TCA_HW_OFFLOAD
+	TCA_INGRESS_BLOCK     = C.TCA_INGRESS_BLOCK
+	TCA_EGRESS_BLOCK      = C.TCA_EGRESS_BLOCK
+	TCA_DUMP_FLAGS        = C.TCA_DUMP_FLAGS
+	TCA_EXT_WARN_MSG      = C.TCA_EXT_WARN_MSG
 	RTNLGRP_NONE          = C.RTNLGRP_NONE
 	RTNLGRP_LINK          = C.RTNLGRP_LINK
 	RTNLGRP_NOTIFY        = C.RTNLGRP_NOTIFY
@@ -3141,6 +3173,18 @@ const (
 	RTNLGRP_IPV6_MROUTE_R = C.RTNLGRP_IPV6_MROUTE_R
 	RTNLGRP_NEXTHOP       = C.RTNLGRP_NEXTHOP
 	RTNLGRP_BRVLAN        = C.RTNLGRP_BRVLAN
+	RTNLGRP_MCTP_IFADDR   = C.RTNLGRP_MCTP_IFADDR
+	RTNLGRP_TUNNEL        = C.RTNLGRP_TUNNEL
+	RTNLGRP_STATS         = C.RTNLGRP_STATS
+	RTNLGRP_IPV4_MCADDR   = C.RTNLGRP_IPV4_MCADDR
+	RTNLGRP_IPV6_MCADDR   = C.RTNLGRP_IPV6_MCADDR
+	RTNLGRP_IPV6_ACADDR   = C.RTNLGRP_IPV6_ACADDR
+	TCA_ROOT_UNSPEC       = C.TCA_ROOT_UNSPEC
+	TCA_ROOT_TAB          = C.TCA_ROOT_TAB
+	TCA_ROOT_FLAGS        = C.TCA_ROOT_FLAGS
+	TCA_ROOT_COUNT        = C.TCA_ROOT_COUNT
+	TCA_ROOT_TIME_DELTA   = C.TCA_ROOT_TIME_DELTA
+	TCA_ROOT_EXT_WARN_MSG = C.TCA_ROOT_EXT_WARN_MSG
 )
 
 // Capabilities
@@ -3556,7 +3600,11 @@ type FsverityEnableArg C.struct_fsverity_enable_arg
 
 type Nhmsg C.struct_nhmsg
 
+const SizeofNhmsg = C.sizeof_struct_nhmsg
+
 type NexthopGrp C.struct_nexthop_grp
+
+const SizeofNexthopGrp = C.sizeof_struct_nexthop_grp
 
 const (
 	NHA_UNSPEC     = C.NHA_UNSPEC
@@ -3823,6 +3871,15 @@ const (
 	ETHTOOL_MSG_PSE_GET                       = C.ETHTOOL_MSG_PSE_GET
 	ETHTOOL_MSG_PSE_SET                       = C.ETHTOOL_MSG_PSE_SET
 	ETHTOOL_MSG_RSS_GET                       = C.ETHTOOL_MSG_RSS_GET
+	ETHTOOL_MSG_PLCA_GET_CFG                  = C.ETHTOOL_MSG_PLCA_GET_CFG
+	ETHTOOL_MSG_PLCA_SET_CFG                  = C.ETHTOOL_MSG_PLCA_SET_CFG
+	ETHTOOL_MSG_PLCA_GET_STATUS               = C.ETHTOOL_MSG_PLCA_GET_STATUS
+	ETHTOOL_MSG_MM_GET                        = C.ETHTOOL_MSG_MM_GET
+	ETHTOOL_MSG_MM_SET                        = C.ETHTOOL_MSG_MM_SET
+	ETHTOOL_MSG_MODULE_FW_FLASH_ACT           = C.ETHTOOL_MSG_MODULE_FW_FLASH_ACT
+	ETHTOOL_MSG_PHY_GET                       = C.ETHTOOL_MSG_PHY_GET
+	ETHTOOL_MSG_TSCONFIG_GET                  = C.ETHTOOL_MSG_TSCONFIG_GET
+	ETHTOOL_MSG_TSCONFIG_SET                  = C.ETHTOOL_MSG_TSCONFIG_SET
 	ETHTOOL_MSG_USER_MAX                      = C.ETHTOOL_MSG_USER_MAX
 	ETHTOOL_MSG_KERNEL_NONE                   = C.ETHTOOL_MSG_KERNEL_NONE
 	ETHTOOL_MSG_STRSET_GET_REPLY              = C.ETHTOOL_MSG_STRSET_GET_REPLY
@@ -3863,6 +3920,16 @@ const (
 	ETHTOOL_MSG_MODULE_NTF                    = C.ETHTOOL_MSG_MODULE_NTF
 	ETHTOOL_MSG_PSE_GET_REPLY                 = C.ETHTOOL_MSG_PSE_GET_REPLY
 	ETHTOOL_MSG_RSS_GET_REPLY                 = C.ETHTOOL_MSG_RSS_GET_REPLY
+	ETHTOOL_MSG_PLCA_GET_CFG_REPLY            = C.ETHTOOL_MSG_PLCA_GET_CFG_REPLY
+	ETHTOOL_MSG_PLCA_GET_STATUS_REPLY         = C.ETHTOOL_MSG_PLCA_GET_STATUS_REPLY
+	ETHTOOL_MSG_PLCA_NTF                      = C.ETHTOOL_MSG_PLCA_NTF
+	ETHTOOL_MSG_MM_GET_REPLY                  = C.ETHTOOL_MSG_MM_GET_REPLY
+	ETHTOOL_MSG_MM_NTF                        = C.ETHTOOL_MSG_MM_NTF
+	ETHTOOL_MSG_MODULE_FW_FLASH_NTF           = C.ETHTOOL_MSG_MODULE_FW_FLASH_NTF
+	ETHTOOL_MSG_PHY_GET_REPLY                 = C.ETHTOOL_MSG_PHY_GET_REPLY
+	ETHTOOL_MSG_PHY_NTF                       = C.ETHTOOL_MSG_PHY_NTF
+	ETHTOOL_MSG_TSCONFIG_GET_REPLY            = C.ETHTOOL_MSG_TSCONFIG_GET_REPLY
+	ETHTOOL_MSG_TSCONFIG_SET_REPLY            = C.ETHTOOL_MSG_TSCONFIG_SET_REPLY
 	ETHTOOL_MSG_KERNEL_MAX                    = C.ETHTOOL_MSG_KERNEL_MAX
 	ETHTOOL_FLAG_COMPACT_BITSETS              = C.ETHTOOL_FLAG_COMPACT_BITSETS
 	ETHTOOL_FLAG_OMIT_REPLY                   = C.ETHTOOL_FLAG_OMIT_REPLY
@@ -3970,6 +4037,11 @@ const (
 	ETHTOOL_A_RINGS_TCP_DATA_SPLIT            = C.ETHTOOL_A_RINGS_TCP_DATA_SPLIT
 	ETHTOOL_A_RINGS_CQE_SIZE                  = C.ETHTOOL_A_RINGS_CQE_SIZE
 	ETHTOOL_A_RINGS_TX_PUSH                   = C.ETHTOOL_A_RINGS_TX_PUSH
+	ETHTOOL_A_RINGS_RX_PUSH                   = C.ETHTOOL_A_RINGS_RX_PUSH
+	ETHTOOL_A_RINGS_TX_PUSH_BUF_LEN           = C.ETHTOOL_A_RINGS_TX_PUSH_BUF_LEN
+	ETHTOOL_A_RINGS_TX_PUSH_BUF_LEN_MAX       = C.ETHTOOL_A_RINGS_TX_PUSH_BUF_LEN_MAX
+	ETHTOOL_A_RINGS_HDS_THRESH                = C.ETHTOOL_A_RINGS_HDS_THRESH
+	ETHTOOL_A_RINGS_HDS_THRESH_MAX            = C.ETHTOOL_A_RINGS_HDS_THRESH_MAX
 	ETHTOOL_A_RINGS_MAX                       = C.ETHTOOL_A_RINGS_MAX
 	ETHTOOL_A_CHANNELS_UNSPEC                 = C.ETHTOOL_A_CHANNELS_UNSPEC
 	ETHTOOL_A_CHANNELS_HEADER                 = C.ETHTOOL_A_CHANNELS_HEADER
@@ -4036,6 +4108,8 @@ const (
 	ETHTOOL_A_TSINFO_TX_TYPES                 = C.ETHTOOL_A_TSINFO_TX_TYPES
 	ETHTOOL_A_TSINFO_RX_FILTERS               = C.ETHTOOL_A_TSINFO_RX_FILTERS
 	ETHTOOL_A_TSINFO_PHC_INDEX                = C.ETHTOOL_A_TSINFO_PHC_INDEX
+	ETHTOOL_A_TSINFO_STATS                    = C.ETHTOOL_A_TSINFO_STATS
+	ETHTOOL_A_TSINFO_HWTSTAMP_PROVIDER        = C.ETHTOOL_A_TSINFO_HWTSTAMP_PROVIDER
 	ETHTOOL_A_TSINFO_MAX                      = C.ETHTOOL_A_TSINFO_MAX
 	ETHTOOL_A_CABLE_TEST_UNSPEC               = C.ETHTOOL_A_CABLE_TEST_UNSPEC
 	ETHTOOL_A_CABLE_TEST_HEADER               = C.ETHTOOL_A_CABLE_TEST_HEADER
@@ -4120,6 +4194,19 @@ const (
 	ETHTOOL_A_TUNNEL_INFO_HEADER              = C.ETHTOOL_A_TUNNEL_INFO_HEADER
 	ETHTOOL_A_TUNNEL_INFO_UDP_PORTS           = C.ETHTOOL_A_TUNNEL_INFO_UDP_PORTS
 	ETHTOOL_A_TUNNEL_INFO_MAX                 = C.ETHTOOL_A_TUNNEL_INFO_MAX
+)
+
+const (
+	TCP_V4_FLOW    = C.TCP_V4_FLOW
+	UDP_V4_FLOW    = C.UDP_V4_FLOW
+	TCP_V6_FLOW    = C.TCP_V6_FLOW
+	UDP_V6_FLOW    = C.UDP_V6_FLOW
+	ESP_V4_FLOW    = C.ESP_V4_FLOW
+	ESP_V6_FLOW    = C.ESP_V6_FLOW
+	IP_USER_FLOW   = C.IP_USER_FLOW
+	IPV6_USER_FLOW = C.IPV6_USER_FLOW
+	IPV6_FLOW      = C.IPV6_FLOW
+	ETHER_FLOW     = C.ETHER_FLOW
 )
 
 const SPEED_UNKNOWN = C.SPEED_UNKNOWN
@@ -4490,6 +4577,7 @@ const (
 	NL80211_ATTR_AKM_SUITES                                 = C.NL80211_ATTR_AKM_SUITES
 	NL80211_ATTR_AP_ISOLATE                                 = C.NL80211_ATTR_AP_ISOLATE
 	NL80211_ATTR_AP_SETTINGS_FLAGS                          = C.NL80211_ATTR_AP_SETTINGS_FLAGS
+	NL80211_ATTR_ASSOC_SPP_AMSDU                            = C.NL80211_ATTR_ASSOC_SPP_AMSDU
 	NL80211_ATTR_AUTH_DATA                                  = C.NL80211_ATTR_AUTH_DATA
 	NL80211_ATTR_AUTH_TYPE                                  = C.NL80211_ATTR_AUTH_TYPE
 	NL80211_ATTR_BANDS                                      = C.NL80211_ATTR_BANDS
@@ -4500,6 +4588,7 @@ const (
 	NL80211_ATTR_BSS_BASIC_RATES                            = C.NL80211_ATTR_BSS_BASIC_RATES
 	NL80211_ATTR_BSS                                        = C.NL80211_ATTR_BSS
 	NL80211_ATTR_BSS_CTS_PROT                               = C.NL80211_ATTR_BSS_CTS_PROT
+	NL80211_ATTR_BSS_DUMP_INCLUDE_USE_DATA                  = C.NL80211_ATTR_BSS_DUMP_INCLUDE_USE_DATA
 	NL80211_ATTR_BSS_HT_OPMODE                              = C.NL80211_ATTR_BSS_HT_OPMODE
 	NL80211_ATTR_BSSID                                      = C.NL80211_ATTR_BSSID
 	NL80211_ATTR_BSS_SELECT                                 = C.NL80211_ATTR_BSS_SELECT
@@ -4559,6 +4648,7 @@ const (
 	NL80211_ATTR_DTIM_PERIOD                                = C.NL80211_ATTR_DTIM_PERIOD
 	NL80211_ATTR_DURATION                                   = C.NL80211_ATTR_DURATION
 	NL80211_ATTR_EHT_CAPABILITY                             = C.NL80211_ATTR_EHT_CAPABILITY
+	NL80211_ATTR_EMA_RNR_ELEMS                              = C.NL80211_ATTR_EMA_RNR_ELEMS
 	NL80211_ATTR_EML_CAPABILITY                             = C.NL80211_ATTR_EML_CAPABILITY
 	NL80211_ATTR_EXT_CAPA                                   = C.NL80211_ATTR_EXT_CAPA
 	NL80211_ATTR_EXT_CAPA_MASK                              = C.NL80211_ATTR_EXT_CAPA_MASK
@@ -4594,6 +4684,7 @@ const (
 	NL80211_ATTR_HIDDEN_SSID                                = C.NL80211_ATTR_HIDDEN_SSID
 	NL80211_ATTR_HT_CAPABILITY                              = C.NL80211_ATTR_HT_CAPABILITY
 	NL80211_ATTR_HT_CAPABILITY_MASK                         = C.NL80211_ATTR_HT_CAPABILITY_MASK
+	NL80211_ATTR_HW_TIMESTAMP_ENABLED                       = C.NL80211_ATTR_HW_TIMESTAMP_ENABLED
 	NL80211_ATTR_IE_ASSOC_RESP                              = C.NL80211_ATTR_IE_ASSOC_RESP
 	NL80211_ATTR_IE                                         = C.NL80211_ATTR_IE
 	NL80211_ATTR_IE_PROBE_RESP                              = C.NL80211_ATTR_IE_PROBE_RESP
@@ -4627,6 +4718,7 @@ const (
 	NL80211_ATTR_MAX                                        = C.NL80211_ATTR_MAX
 	NL80211_ATTR_MAX_CRIT_PROT_DURATION                     = C.NL80211_ATTR_MAX_CRIT_PROT_DURATION
 	NL80211_ATTR_MAX_CSA_COUNTERS                           = C.NL80211_ATTR_MAX_CSA_COUNTERS
+	NL80211_ATTR_MAX_HW_TIMESTAMP_PEERS                     = C.NL80211_ATTR_MAX_HW_TIMESTAMP_PEERS
 	NL80211_ATTR_MAX_MATCH_SETS                             = C.NL80211_ATTR_MAX_MATCH_SETS
 	NL80211_ATTR_MAX_NUM_AKM_SUITES                         = C.NL80211_ATTR_MAX_NUM_AKM_SUITES
 	NL80211_ATTR_MAX_NUM_PMKIDS                             = C.NL80211_ATTR_MAX_NUM_PMKIDS
@@ -4651,9 +4743,12 @@ const (
 	NL80211_ATTR_MGMT_SUBTYPE                               = C.NL80211_ATTR_MGMT_SUBTYPE
 	NL80211_ATTR_MLD_ADDR                                   = C.NL80211_ATTR_MLD_ADDR
 	NL80211_ATTR_MLD_CAPA_AND_OPS                           = C.NL80211_ATTR_MLD_CAPA_AND_OPS
+	NL80211_ATTR_MLO_LINK_DISABLED                          = C.NL80211_ATTR_MLO_LINK_DISABLED
 	NL80211_ATTR_MLO_LINK_ID                                = C.NL80211_ATTR_MLO_LINK_ID
 	NL80211_ATTR_MLO_LINKS                                  = C.NL80211_ATTR_MLO_LINKS
 	NL80211_ATTR_MLO_SUPPORT                                = C.NL80211_ATTR_MLO_SUPPORT
+	NL80211_ATTR_MLO_TTLM_DLINK                             = C.NL80211_ATTR_MLO_TTLM_DLINK
+	NL80211_ATTR_MLO_TTLM_ULINK                             = C.NL80211_ATTR_MLO_TTLM_ULINK
 	NL80211_ATTR_MNTR_FLAGS                                 = C.NL80211_ATTR_MNTR_FLAGS
 	NL80211_ATTR_MPATH_INFO                                 = C.NL80211_ATTR_MPATH_INFO
 	NL80211_ATTR_MPATH_NEXT_HOP                             = C.NL80211_ATTR_MPATH_NEXT_HOP
@@ -4686,12 +4781,14 @@ const (
 	NL80211_ATTR_PORT_AUTHORIZED                            = C.NL80211_ATTR_PORT_AUTHORIZED
 	NL80211_ATTR_POWER_RULE_MAX_ANT_GAIN                    = C.NL80211_ATTR_POWER_RULE_MAX_ANT_GAIN
 	NL80211_ATTR_POWER_RULE_MAX_EIRP                        = C.NL80211_ATTR_POWER_RULE_MAX_EIRP
+	NL80211_ATTR_POWER_RULE_PSD                             = C.NL80211_ATTR_POWER_RULE_PSD
 	NL80211_ATTR_PREV_BSSID                                 = C.NL80211_ATTR_PREV_BSSID
 	NL80211_ATTR_PRIVACY                                    = C.NL80211_ATTR_PRIVACY
 	NL80211_ATTR_PROBE_RESP                                 = C.NL80211_ATTR_PROBE_RESP
 	NL80211_ATTR_PROBE_RESP_OFFLOAD                         = C.NL80211_ATTR_PROBE_RESP_OFFLOAD
 	NL80211_ATTR_PROTOCOL_FEATURES                          = C.NL80211_ATTR_PROTOCOL_FEATURES
 	NL80211_ATTR_PS_STATE                                   = C.NL80211_ATTR_PS_STATE
+	NL80211_ATTR_PUNCT_BITMAP                               = C.NL80211_ATTR_PUNCT_BITMAP
 	NL80211_ATTR_QOS_MAP                                    = C.NL80211_ATTR_QOS_MAP
 	NL80211_ATTR_RADAR_BACKGROUND                           = C.NL80211_ATTR_RADAR_BACKGROUND
 	NL80211_ATTR_RADAR_EVENT                                = C.NL80211_ATTR_RADAR_EVENT
@@ -4820,7 +4917,9 @@ const (
 	NL80211_ATTR_WIPHY_FREQ                                 = C.NL80211_ATTR_WIPHY_FREQ
 	NL80211_ATTR_WIPHY_FREQ_HINT                            = C.NL80211_ATTR_WIPHY_FREQ_HINT
 	NL80211_ATTR_WIPHY_FREQ_OFFSET                          = C.NL80211_ATTR_WIPHY_FREQ_OFFSET
+	NL80211_ATTR_WIPHY_INTERFACE_COMBINATIONS               = C.NL80211_ATTR_WIPHY_INTERFACE_COMBINATIONS
 	NL80211_ATTR_WIPHY_NAME                                 = C.NL80211_ATTR_WIPHY_NAME
+	NL80211_ATTR_WIPHY_RADIOS                               = C.NL80211_ATTR_WIPHY_RADIOS
 	NL80211_ATTR_WIPHY_RETRY_LONG                           = C.NL80211_ATTR_WIPHY_RETRY_LONG
 	NL80211_ATTR_WIPHY_RETRY_SHORT                          = C.NL80211_ATTR_WIPHY_RETRY_SHORT
 	NL80211_ATTR_WIPHY_RTS_THRESHOLD                        = C.NL80211_ATTR_WIPHY_RTS_THRESHOLD
@@ -4855,6 +4954,8 @@ const (
 	NL80211_BAND_ATTR_IFTYPE_DATA                           = C.NL80211_BAND_ATTR_IFTYPE_DATA
 	NL80211_BAND_ATTR_MAX                                   = C.NL80211_BAND_ATTR_MAX
 	NL80211_BAND_ATTR_RATES                                 = C.NL80211_BAND_ATTR_RATES
+	NL80211_BAND_ATTR_S1G_CAPA                              = C.NL80211_BAND_ATTR_S1G_CAPA
+	NL80211_BAND_ATTR_S1G_MCS_NSS_SET                       = C.NL80211_BAND_ATTR_S1G_MCS_NSS_SET
 	NL80211_BAND_ATTR_VHT_CAPA                              = C.NL80211_BAND_ATTR_VHT_CAPA
 	NL80211_BAND_ATTR_VHT_MCS_SET                           = C.NL80211_BAND_ATTR_VHT_MCS_SET
 	NL80211_BAND_IFTYPE_ATTR_EHT_CAP_MAC                    = C.NL80211_BAND_IFTYPE_ATTR_EHT_CAP_MAC
@@ -4878,6 +4979,10 @@ const (
 	NL80211_BSS_BEACON_INTERVAL                             = C.NL80211_BSS_BEACON_INTERVAL
 	NL80211_BSS_BEACON_TSF                                  = C.NL80211_BSS_BEACON_TSF
 	NL80211_BSS_BSSID                                       = C.NL80211_BSS_BSSID
+	NL80211_BSS_CANNOT_USE_6GHZ_PWR_MISMATCH                = C.NL80211_BSS_CANNOT_USE_6GHZ_PWR_MISMATCH
+	NL80211_BSS_CANNOT_USE_NSTR_NONPRIMARY                  = C.NL80211_BSS_CANNOT_USE_NSTR_NONPRIMARY
+	NL80211_BSS_CANNOT_USE_REASONS                          = C.NL80211_BSS_CANNOT_USE_REASONS
+	NL80211_BSS_CANNOT_USE_UHB_PWR_MISMATCH                 = C.NL80211_BSS_CANNOT_USE_UHB_PWR_MISMATCH
 	NL80211_BSS_CAPABILITY                                  = C.NL80211_BSS_CAPABILITY
 	NL80211_BSS_CHAIN_SIGNAL                                = C.NL80211_BSS_CHAIN_SIGNAL
 	NL80211_BSS_CHAN_WIDTH_10                               = C.NL80211_BSS_CHAN_WIDTH_10
@@ -4909,6 +5014,9 @@ const (
 	NL80211_BSS_STATUS                                      = C.NL80211_BSS_STATUS
 	NL80211_BSS_STATUS_IBSS_JOINED                          = C.NL80211_BSS_STATUS_IBSS_JOINED
 	NL80211_BSS_TSF                                         = C.NL80211_BSS_TSF
+	NL80211_BSS_USE_FOR                                     = C.NL80211_BSS_USE_FOR
+	NL80211_BSS_USE_FOR_MLD_LINK                            = C.NL80211_BSS_USE_FOR_MLD_LINK
+	NL80211_BSS_USE_FOR_NORMAL                              = C.NL80211_BSS_USE_FOR_NORMAL
 	NL80211_CHAN_HT20                                       = C.NL80211_CHAN_HT20
 	NL80211_CHAN_HT40MINUS                                  = C.NL80211_CHAN_HT40MINUS
 	NL80211_CHAN_HT40PLUS                                   = C.NL80211_CHAN_HT40PLUS
@@ -4994,6 +5102,7 @@ const (
 	NL80211_CMD_LEAVE_IBSS                                  = C.NL80211_CMD_LEAVE_IBSS
 	NL80211_CMD_LEAVE_MESH                                  = C.NL80211_CMD_LEAVE_MESH
 	NL80211_CMD_LEAVE_OCB                                   = C.NL80211_CMD_LEAVE_OCB
+	NL80211_CMD_LINKS_REMOVED                               = C.NL80211_CMD_LINKS_REMOVED
 	NL80211_CMD_MAX                                         = C.NL80211_CMD_MAX
 	NL80211_CMD_MICHAEL_MIC_FAILURE                         = C.NL80211_CMD_MICHAEL_MIC_FAILURE
 	NL80211_CMD_MODIFY_LINK_STA                             = C.NL80211_CMD_MODIFY_LINK_STA
@@ -5038,6 +5147,7 @@ const (
 	NL80211_CMD_SET_COALESCE                                = C.NL80211_CMD_SET_COALESCE
 	NL80211_CMD_SET_CQM                                     = C.NL80211_CMD_SET_CQM
 	NL80211_CMD_SET_FILS_AAD                                = C.NL80211_CMD_SET_FILS_AAD
+	NL80211_CMD_SET_HW_TIMESTAMP                            = C.NL80211_CMD_SET_HW_TIMESTAMP
 	NL80211_CMD_SET_INTERFACE                               = C.NL80211_CMD_SET_INTERFACE
 	NL80211_CMD_SET_KEY                                     = C.NL80211_CMD_SET_KEY
 	NL80211_CMD_SET_MAC_ACL                                 = C.NL80211_CMD_SET_MAC_ACL
@@ -5057,6 +5167,7 @@ const (
 	NL80211_CMD_SET_SAR_SPECS                               = C.NL80211_CMD_SET_SAR_SPECS
 	NL80211_CMD_SET_STATION                                 = C.NL80211_CMD_SET_STATION
 	NL80211_CMD_SET_TID_CONFIG                              = C.NL80211_CMD_SET_TID_CONFIG
+	NL80211_CMD_SET_TID_TO_LINK_MAPPING                     = C.NL80211_CMD_SET_TID_TO_LINK_MAPPING
 	NL80211_CMD_SET_TX_BITRATE_MASK                         = C.NL80211_CMD_SET_TX_BITRATE_MASK
 	NL80211_CMD_SET_WDS_PEER                                = C.NL80211_CMD_SET_WDS_PEER
 	NL80211_CMD_SET_WIPHY                                   = C.NL80211_CMD_SET_WIPHY
@@ -5124,6 +5235,7 @@ const (
 	NL80211_EXT_FEATURE_AIRTIME_FAIRNESS                    = C.NL80211_EXT_FEATURE_AIRTIME_FAIRNESS
 	NL80211_EXT_FEATURE_AP_PMKSA_CACHING                    = C.NL80211_EXT_FEATURE_AP_PMKSA_CACHING
 	NL80211_EXT_FEATURE_AQL                                 = C.NL80211_EXT_FEATURE_AQL
+	NL80211_EXT_FEATURE_AUTH_AND_DEAUTH_RANDOM_TA           = C.NL80211_EXT_FEATURE_AUTH_AND_DEAUTH_RANDOM_TA
 	NL80211_EXT_FEATURE_BEACON_PROTECTION_CLIENT            = C.NL80211_EXT_FEATURE_BEACON_PROTECTION_CLIENT
 	NL80211_EXT_FEATURE_BEACON_PROTECTION                   = C.NL80211_EXT_FEATURE_BEACON_PROTECTION
 	NL80211_EXT_FEATURE_BEACON_RATE_HE                      = C.NL80211_EXT_FEATURE_BEACON_RATE_HE
@@ -5139,6 +5251,7 @@ const (
 	NL80211_EXT_FEATURE_CQM_RSSI_LIST                       = C.NL80211_EXT_FEATURE_CQM_RSSI_LIST
 	NL80211_EXT_FEATURE_DATA_ACK_SIGNAL_SUPPORT             = C.NL80211_EXT_FEATURE_DATA_ACK_SIGNAL_SUPPORT
 	NL80211_EXT_FEATURE_DEL_IBSS_STA                        = C.NL80211_EXT_FEATURE_DEL_IBSS_STA
+	NL80211_EXT_FEATURE_DFS_CONCURRENT                      = C.NL80211_EXT_FEATURE_DFS_CONCURRENT
 	NL80211_EXT_FEATURE_DFS_OFFLOAD                         = C.NL80211_EXT_FEATURE_DFS_OFFLOAD
 	NL80211_EXT_FEATURE_ENABLE_FTM_RESPONDER                = C.NL80211_EXT_FEATURE_ENABLE_FTM_RESPONDER
 	NL80211_EXT_FEATURE_EXT_KEY_ID                          = C.NL80211_EXT_FEATURE_EXT_KEY_ID
@@ -5158,9 +5271,12 @@ const (
 	NL80211_EXT_FEATURE_OCE_PROBE_REQ_DEFERRAL_SUPPRESSION  = C.NL80211_EXT_FEATURE_OCE_PROBE_REQ_DEFERRAL_SUPPRESSION
 	NL80211_EXT_FEATURE_OCE_PROBE_REQ_HIGH_TX_RATE          = C.NL80211_EXT_FEATURE_OCE_PROBE_REQ_HIGH_TX_RATE
 	NL80211_EXT_FEATURE_OPERATING_CHANNEL_VALIDATION        = C.NL80211_EXT_FEATURE_OPERATING_CHANNEL_VALIDATION
+	NL80211_EXT_FEATURE_OWE_OFFLOAD_AP                      = C.NL80211_EXT_FEATURE_OWE_OFFLOAD_AP
+	NL80211_EXT_FEATURE_OWE_OFFLOAD                         = C.NL80211_EXT_FEATURE_OWE_OFFLOAD
 	NL80211_EXT_FEATURE_POWERED_ADDR_CHANGE                 = C.NL80211_EXT_FEATURE_POWERED_ADDR_CHANGE
 	NL80211_EXT_FEATURE_PROTECTED_TWT                       = C.NL80211_EXT_FEATURE_PROTECTED_TWT
 	NL80211_EXT_FEATURE_PROT_RANGE_NEGO_AND_MEASURE         = C.NL80211_EXT_FEATURE_PROT_RANGE_NEGO_AND_MEASURE
+	NL80211_EXT_FEATURE_PUNCT                               = C.NL80211_EXT_FEATURE_PUNCT
 	NL80211_EXT_FEATURE_RADAR_BACKGROUND                    = C.NL80211_EXT_FEATURE_RADAR_BACKGROUND
 	NL80211_EXT_FEATURE_RRM                                 = C.NL80211_EXT_FEATURE_RRM
 	NL80211_EXT_FEATURE_SAE_OFFLOAD_AP                      = C.NL80211_EXT_FEATURE_SAE_OFFLOAD_AP
@@ -5172,8 +5288,10 @@ const (
 	NL80211_EXT_FEATURE_SCHED_SCAN_BAND_SPECIFIC_RSSI_THOLD = C.NL80211_EXT_FEATURE_SCHED_SCAN_BAND_SPECIFIC_RSSI_THOLD
 	NL80211_EXT_FEATURE_SCHED_SCAN_RELATIVE_RSSI            = C.NL80211_EXT_FEATURE_SCHED_SCAN_RELATIVE_RSSI
 	NL80211_EXT_FEATURE_SECURE_LTF                          = C.NL80211_EXT_FEATURE_SECURE_LTF
+	NL80211_EXT_FEATURE_SECURE_NAN                          = C.NL80211_EXT_FEATURE_SECURE_NAN
 	NL80211_EXT_FEATURE_SECURE_RTT                          = C.NL80211_EXT_FEATURE_SECURE_RTT
 	NL80211_EXT_FEATURE_SET_SCAN_DWELL                      = C.NL80211_EXT_FEATURE_SET_SCAN_DWELL
+	NL80211_EXT_FEATURE_SPP_AMSDU_SUPPORT                   = C.NL80211_EXT_FEATURE_SPP_AMSDU_SUPPORT
 	NL80211_EXT_FEATURE_STA_TX_PWR                          = C.NL80211_EXT_FEATURE_STA_TX_PWR
 	NL80211_EXT_FEATURE_TXQS                                = C.NL80211_EXT_FEATURE_TXQS
 	NL80211_EXT_FEATURE_UNSOL_BCAST_PROBE_RESP              = C.NL80211_EXT_FEATURE_UNSOL_BCAST_PROBE_RESP
@@ -5220,7 +5338,10 @@ const (
 	NL80211_FREQUENCY_ATTR_2MHZ                             = C.NL80211_FREQUENCY_ATTR_2MHZ
 	NL80211_FREQUENCY_ATTR_4MHZ                             = C.NL80211_FREQUENCY_ATTR_4MHZ
 	NL80211_FREQUENCY_ATTR_8MHZ                             = C.NL80211_FREQUENCY_ATTR_8MHZ
+	NL80211_FREQUENCY_ATTR_ALLOW_6GHZ_VLP_AP                = C.NL80211_FREQUENCY_ATTR_ALLOW_6GHZ_VLP_AP
+	NL80211_FREQUENCY_ATTR_CAN_MONITOR                      = C.NL80211_FREQUENCY_ATTR_CAN_MONITOR
 	NL80211_FREQUENCY_ATTR_DFS_CAC_TIME                     = C.NL80211_FREQUENCY_ATTR_DFS_CAC_TIME
+	NL80211_FREQUENCY_ATTR_DFS_CONCURRENT                   = C.NL80211_FREQUENCY_ATTR_DFS_CONCURRENT
 	NL80211_FREQUENCY_ATTR_DFS_STATE                        = C.NL80211_FREQUENCY_ATTR_DFS_STATE
 	NL80211_FREQUENCY_ATTR_DFS_TIME                         = C.NL80211_FREQUENCY_ATTR_DFS_TIME
 	NL80211_FREQUENCY_ATTR_DISABLED                         = C.NL80211_FREQUENCY_ATTR_DISABLED
@@ -5234,6 +5355,8 @@ const (
 	NL80211_FREQUENCY_ATTR_NO_160MHZ                        = C.NL80211_FREQUENCY_ATTR_NO_160MHZ
 	NL80211_FREQUENCY_ATTR_NO_20MHZ                         = C.NL80211_FREQUENCY_ATTR_NO_20MHZ
 	NL80211_FREQUENCY_ATTR_NO_320MHZ                        = C.NL80211_FREQUENCY_ATTR_NO_320MHZ
+	NL80211_FREQUENCY_ATTR_NO_6GHZ_AFC_CLIENT               = C.NL80211_FREQUENCY_ATTR_NO_6GHZ_AFC_CLIENT
+	NL80211_FREQUENCY_ATTR_NO_6GHZ_VLP_CLIENT               = C.NL80211_FREQUENCY_ATTR_NO_6GHZ_VLP_CLIENT
 	NL80211_FREQUENCY_ATTR_NO_80MHZ                         = C.NL80211_FREQUENCY_ATTR_NO_80MHZ
 	NL80211_FREQUENCY_ATTR_NO_EHT                           = C.NL80211_FREQUENCY_ATTR_NO_EHT
 	NL80211_FREQUENCY_ATTR_NO_HE                            = C.NL80211_FREQUENCY_ATTR_NO_HE
@@ -5241,8 +5364,11 @@ const (
 	NL80211_FREQUENCY_ATTR_NO_HT40_PLUS                     = C.NL80211_FREQUENCY_ATTR_NO_HT40_PLUS
 	NL80211_FREQUENCY_ATTR_NO_IBSS                          = C.NL80211_FREQUENCY_ATTR_NO_IBSS
 	NL80211_FREQUENCY_ATTR_NO_IR                            = C.NL80211_FREQUENCY_ATTR_NO_IR
+	NL80211_FREQUENCY_ATTR_NO_UHB_AFC_CLIENT                = C.NL80211_FREQUENCY_ATTR_NO_UHB_AFC_CLIENT
+	NL80211_FREQUENCY_ATTR_NO_UHB_VLP_CLIENT                = C.NL80211_FREQUENCY_ATTR_NO_UHB_VLP_CLIENT
 	NL80211_FREQUENCY_ATTR_OFFSET                           = C.NL80211_FREQUENCY_ATTR_OFFSET
 	NL80211_FREQUENCY_ATTR_PASSIVE_SCAN                     = C.NL80211_FREQUENCY_ATTR_PASSIVE_SCAN
+	NL80211_FREQUENCY_ATTR_PSD                              = C.NL80211_FREQUENCY_ATTR_PSD
 	NL80211_FREQUENCY_ATTR_RADAR                            = C.NL80211_FREQUENCY_ATTR_RADAR
 	NL80211_FREQUENCY_ATTR_WMM                              = C.NL80211_FREQUENCY_ATTR_WMM
 	NL80211_FTM_RESP_ATTR_CIVICLOC                          = C.NL80211_FTM_RESP_ATTR_CIVICLOC
@@ -5307,6 +5433,7 @@ const (
 	NL80211_IFTYPE_STATION                                  = C.NL80211_IFTYPE_STATION
 	NL80211_IFTYPE_UNSPECIFIED                              = C.NL80211_IFTYPE_UNSPECIFIED
 	NL80211_IFTYPE_WDS                                      = C.NL80211_IFTYPE_WDS
+	NL80211_KCK_EXT_LEN_32                                  = C.NL80211_KCK_EXT_LEN_32
 	NL80211_KCK_EXT_LEN                                     = C.NL80211_KCK_EXT_LEN
 	NL80211_KCK_LEN                                         = C.NL80211_KCK_LEN
 	NL80211_KEK_EXT_LEN                                     = C.NL80211_KEK_EXT_LEN
@@ -5335,6 +5462,7 @@ const (
 	NL80211_MAX_SUPP_HT_RATES                               = C.NL80211_MAX_SUPP_HT_RATES
 	NL80211_MAX_SUPP_RATES                                  = C.NL80211_MAX_SUPP_RATES
 	NL80211_MAX_SUPP_REG_RULES                              = C.NL80211_MAX_SUPP_REG_RULES
+	NL80211_MAX_SUPP_SELECTORS                              = C.NL80211_MAX_SUPP_SELECTORS
 	NL80211_MBSSID_CONFIG_ATTR_EMA                          = C.NL80211_MBSSID_CONFIG_ATTR_EMA
 	NL80211_MBSSID_CONFIG_ATTR_INDEX                        = C.NL80211_MBSSID_CONFIG_ATTR_INDEX
 	NL80211_MBSSID_CONFIG_ATTR_MAX                          = C.NL80211_MBSSID_CONFIG_ATTR_MAX
@@ -5580,11 +5708,16 @@ const (
 	NL80211_RADAR_PRE_CAC_EXPIRED                           = C.NL80211_RADAR_PRE_CAC_EXPIRED
 	NL80211_RATE_INFO_10_MHZ_WIDTH                          = C.NL80211_RATE_INFO_10_MHZ_WIDTH
 	NL80211_RATE_INFO_160_MHZ_WIDTH                         = C.NL80211_RATE_INFO_160_MHZ_WIDTH
+	NL80211_RATE_INFO_16_MHZ_WIDTH                          = C.NL80211_RATE_INFO_16_MHZ_WIDTH
+	NL80211_RATE_INFO_1_MHZ_WIDTH                           = C.NL80211_RATE_INFO_1_MHZ_WIDTH
+	NL80211_RATE_INFO_2_MHZ_WIDTH                           = C.NL80211_RATE_INFO_2_MHZ_WIDTH
 	NL80211_RATE_INFO_320_MHZ_WIDTH                         = C.NL80211_RATE_INFO_320_MHZ_WIDTH
 	NL80211_RATE_INFO_40_MHZ_WIDTH                          = C.NL80211_RATE_INFO_40_MHZ_WIDTH
+	NL80211_RATE_INFO_4_MHZ_WIDTH                           = C.NL80211_RATE_INFO_4_MHZ_WIDTH
 	NL80211_RATE_INFO_5_MHZ_WIDTH                           = C.NL80211_RATE_INFO_5_MHZ_WIDTH
 	NL80211_RATE_INFO_80_MHZ_WIDTH                          = C.NL80211_RATE_INFO_80_MHZ_WIDTH
 	NL80211_RATE_INFO_80P80_MHZ_WIDTH                       = C.NL80211_RATE_INFO_80P80_MHZ_WIDTH
+	NL80211_RATE_INFO_8_MHZ_WIDTH                           = C.NL80211_RATE_INFO_8_MHZ_WIDTH
 	NL80211_RATE_INFO_BITRATE32                             = C.NL80211_RATE_INFO_BITRATE32
 	NL80211_RATE_INFO_BITRATE                               = C.NL80211_RATE_INFO_BITRATE
 	NL80211_RATE_INFO_EHT_GI_0_8                            = C.NL80211_RATE_INFO_EHT_GI_0_8
@@ -5630,6 +5763,8 @@ const (
 	NL80211_RATE_INFO_HE_RU_ALLOC                           = C.NL80211_RATE_INFO_HE_RU_ALLOC
 	NL80211_RATE_INFO_MAX                                   = C.NL80211_RATE_INFO_MAX
 	NL80211_RATE_INFO_MCS                                   = C.NL80211_RATE_INFO_MCS
+	NL80211_RATE_INFO_S1G_MCS                               = C.NL80211_RATE_INFO_S1G_MCS
+	NL80211_RATE_INFO_S1G_NSS                               = C.NL80211_RATE_INFO_S1G_NSS
 	NL80211_RATE_INFO_SHORT_GI                              = C.NL80211_RATE_INFO_SHORT_GI
 	NL80211_RATE_INFO_VHT_MCS                               = C.NL80211_RATE_INFO_VHT_MCS
 	NL80211_RATE_INFO_VHT_NSS                               = C.NL80211_RATE_INFO_VHT_NSS
@@ -5647,14 +5782,19 @@ const (
 	NL80211_REKEY_DATA_KEK                                  = C.NL80211_REKEY_DATA_KEK
 	NL80211_REKEY_DATA_REPLAY_CTR                           = C.NL80211_REKEY_DATA_REPLAY_CTR
 	NL80211_REPLAY_CTR_LEN                                  = C.NL80211_REPLAY_CTR_LEN
+	NL80211_RRF_ALLOW_6GHZ_VLP_AP                           = C.NL80211_RRF_ALLOW_6GHZ_VLP_AP
 	NL80211_RRF_AUTO_BW                                     = C.NL80211_RRF_AUTO_BW
 	NL80211_RRF_DFS                                         = C.NL80211_RRF_DFS
+	NL80211_RRF_DFS_CONCURRENT                              = C.NL80211_RRF_DFS_CONCURRENT
 	NL80211_RRF_GO_CONCURRENT                               = C.NL80211_RRF_GO_CONCURRENT
 	NL80211_RRF_IR_CONCURRENT                               = C.NL80211_RRF_IR_CONCURRENT
 	NL80211_RRF_NO_160MHZ                                   = C.NL80211_RRF_NO_160MHZ
 	NL80211_RRF_NO_320MHZ                                   = C.NL80211_RRF_NO_320MHZ
+	NL80211_RRF_NO_6GHZ_AFC_CLIENT                          = C.NL80211_RRF_NO_6GHZ_AFC_CLIENT
+	NL80211_RRF_NO_6GHZ_VLP_CLIENT                          = C.NL80211_RRF_NO_6GHZ_VLP_CLIENT
 	NL80211_RRF_NO_80MHZ                                    = C.NL80211_RRF_NO_80MHZ
 	NL80211_RRF_NO_CCK                                      = C.NL80211_RRF_NO_CCK
+	NL80211_RRF_NO_EHT                                      = C.NL80211_RRF_NO_EHT
 	NL80211_RRF_NO_HE                                       = C.NL80211_RRF_NO_HE
 	NL80211_RRF_NO_HT40                                     = C.NL80211_RRF_NO_HT40
 	NL80211_RRF_NO_HT40MINUS                                = C.NL80211_RRF_NO_HT40MINUS
@@ -5665,7 +5805,10 @@ const (
 	NL80211_RRF_NO_IR                                       = C.NL80211_RRF_NO_IR
 	NL80211_RRF_NO_OFDM                                     = C.NL80211_RRF_NO_OFDM
 	NL80211_RRF_NO_OUTDOOR                                  = C.NL80211_RRF_NO_OUTDOOR
+	NL80211_RRF_NO_UHB_AFC_CLIENT                           = C.NL80211_RRF_NO_UHB_AFC_CLIENT
+	NL80211_RRF_NO_UHB_VLP_CLIENT                           = C.NL80211_RRF_NO_UHB_VLP_CLIENT
 	NL80211_RRF_PASSIVE_SCAN                                = C.NL80211_RRF_PASSIVE_SCAN
+	NL80211_RRF_PSD                                         = C.NL80211_RRF_PSD
 	NL80211_RRF_PTMP_ONLY                                   = C.NL80211_RRF_PTMP_ONLY
 	NL80211_RRF_PTP_ONLY                                    = C.NL80211_RRF_PTP_ONLY
 	NL80211_RXMGMT_FLAG_ANSWERED                            = C.NL80211_RXMGMT_FLAG_ANSWERED
@@ -5726,6 +5869,7 @@ const (
 	NL80211_STA_FLAG_MAX_OLD_API                            = C.NL80211_STA_FLAG_MAX_OLD_API
 	NL80211_STA_FLAG_MFP                                    = C.NL80211_STA_FLAG_MFP
 	NL80211_STA_FLAG_SHORT_PREAMBLE                         = C.NL80211_STA_FLAG_SHORT_PREAMBLE
+	NL80211_STA_FLAG_SPP_AMSDU                              = C.NL80211_STA_FLAG_SPP_AMSDU
 	NL80211_STA_FLAG_TDLS_PEER                              = C.NL80211_STA_FLAG_TDLS_PEER
 	NL80211_STA_FLAG_WME                                    = C.NL80211_STA_FLAG_WME
 	NL80211_STA_INFO_ACK_SIGNAL_AVG                         = C.NL80211_STA_INFO_ACK_SIGNAL_AVG
@@ -5884,6 +6028,13 @@ const (
 	NL80211_VHT_CAPABILITY_LEN                              = C.NL80211_VHT_CAPABILITY_LEN
 	NL80211_VHT_NSS_MAX                                     = C.NL80211_VHT_NSS_MAX
 	NL80211_WIPHY_NAME_MAXLEN                               = C.NL80211_WIPHY_NAME_MAXLEN
+	NL80211_WIPHY_RADIO_ATTR_FREQ_RANGE                     = C.NL80211_WIPHY_RADIO_ATTR_FREQ_RANGE
+	NL80211_WIPHY_RADIO_ATTR_INDEX                          = C.NL80211_WIPHY_RADIO_ATTR_INDEX
+	NL80211_WIPHY_RADIO_ATTR_INTERFACE_COMBINATION          = C.NL80211_WIPHY_RADIO_ATTR_INTERFACE_COMBINATION
+	NL80211_WIPHY_RADIO_ATTR_MAX                            = C.NL80211_WIPHY_RADIO_ATTR_MAX
+	NL80211_WIPHY_RADIO_FREQ_ATTR_END                       = C.NL80211_WIPHY_RADIO_FREQ_ATTR_END
+	NL80211_WIPHY_RADIO_FREQ_ATTR_MAX                       = C.NL80211_WIPHY_RADIO_FREQ_ATTR_MAX
+	NL80211_WIPHY_RADIO_FREQ_ATTR_START                     = C.NL80211_WIPHY_RADIO_FREQ_ATTR_START
 	NL80211_WMMR_AIFSN                                      = C.NL80211_WMMR_AIFSN
 	NL80211_WMMR_CW_MAX                                     = C.NL80211_WMMR_CW_MAX
 	NL80211_WMMR_CW_MIN                                     = C.NL80211_WMMR_CW_MIN
@@ -5915,6 +6066,7 @@ const (
 	NL80211_WOWLAN_TRIG_PKT_PATTERN                         = C.NL80211_WOWLAN_TRIG_PKT_PATTERN
 	NL80211_WOWLAN_TRIG_RFKILL_RELEASE                      = C.NL80211_WOWLAN_TRIG_RFKILL_RELEASE
 	NL80211_WOWLAN_TRIG_TCP_CONNECTION                      = C.NL80211_WOWLAN_TRIG_TCP_CONNECTION
+	NL80211_WOWLAN_TRIG_UNPROTECTED_DEAUTH_DISASSOC         = C.NL80211_WOWLAN_TRIG_UNPROTECTED_DEAUTH_DISASSOC
 	NL80211_WOWLAN_TRIG_WAKEUP_PKT_80211                    = C.NL80211_WOWLAN_TRIG_WAKEUP_PKT_80211
 	NL80211_WOWLAN_TRIG_WAKEUP_PKT_80211_LEN                = C.NL80211_WOWLAN_TRIG_WAKEUP_PKT_80211_LEN
 	NL80211_WOWLAN_TRIG_WAKEUP_PKT_8023                     = C.NL80211_WOWLAN_TRIG_WAKEUP_PKT_8023
@@ -6103,3 +6255,34 @@ type SockDiagReq C.struct_sock_diag_req
 
 // Removed in Linux 6.13, kept for backwards compatibility.
 const RTM_NEWNVLAN = 0x70
+
+// Memory policy modes and flags for [SetMemPolicy], as defined in /usr/include/linux/mempolicy.h.
+// Not all can be used, see set_mempolicy(2).
+// Generated by:
+// $ { perl -nlE '/^#define (MPOL_\w+)/ && say "\t$1 = C.$1"' /usr/include/linux/mempolicy.h; perl -nlE '/^\s*(MPOL_\w+)/ && say "\t$1 = C.$1"' /usr/include/linux/mempolicy.h; } | sort | uniq
+const (
+	MPOL_BIND                = C.MPOL_BIND
+	MPOL_DEFAULT             = C.MPOL_DEFAULT
+	MPOL_F_ADDR              = C.MPOL_F_ADDR
+	MPOL_F_MEMS_ALLOWED      = C.MPOL_F_MEMS_ALLOWED
+	MPOL_F_MOF               = C.MPOL_F_MOF
+	MPOL_F_MORON             = C.MPOL_F_MORON
+	MPOL_F_NODE              = C.MPOL_F_NODE
+	MPOL_F_NUMA_BALANCING    = C.MPOL_F_NUMA_BALANCING
+	MPOL_F_RELATIVE_NODES    = C.MPOL_F_RELATIVE_NODES
+	MPOL_F_SHARED            = C.MPOL_F_SHARED
+	MPOL_F_STATIC_NODES      = C.MPOL_F_STATIC_NODES
+	MPOL_INTERLEAVE          = C.MPOL_INTERLEAVE
+	MPOL_LOCAL               = C.MPOL_LOCAL
+	MPOL_MAX                 = C.MPOL_MAX
+	MPOL_MF_INTERNAL         = C.MPOL_MF_INTERNAL
+	MPOL_MF_LAZY             = C.MPOL_MF_LAZY
+	MPOL_MF_MOVE_ALL         = C.MPOL_MF_MOVE_ALL
+	MPOL_MF_MOVE             = C.MPOL_MF_MOVE
+	MPOL_MF_STRICT           = C.MPOL_MF_STRICT
+	MPOL_MF_VALID            = C.MPOL_MF_VALID
+	MPOL_MODE_FLAGS          = C.MPOL_MODE_FLAGS
+	MPOL_PREFERRED           = C.MPOL_PREFERRED
+	MPOL_PREFERRED_MANY      = C.MPOL_PREFERRED_MANY
+	MPOL_WEIGHTED_INTERLEAVE = C.MPOL_WEIGHTED_INTERLEAVE
+)
