@@ -50,7 +50,7 @@ func TestDLLFilenameEscaping(t *testing.T) {
 	}
 }
 
-func TestSyscallXGeneration(t *testing.T) {
+func TestSyscallNGeneration(t *testing.T) {
 	tests := []struct {
 		name        string
 		wantsysfunc string
@@ -58,17 +58,17 @@ func TestSyscallXGeneration(t *testing.T) {
 	}{
 		{
 			name:        "syscall with 2 params",
-			wantsysfunc: "syscall.Syscall",
+			wantsysfunc: "syscall.SyscallN",
 			sig:         "Example(a1 *uint16, a2 *uint16) = ",
 		},
 		{
 			name:        "syscall with 6 params",
-			wantsysfunc: "syscall.Syscall6",
+			wantsysfunc: "syscall.SyscallN",
 			sig:         "Example(a1 *uint, a2 *uint, a3 *uint, a4 *uint, a5 *uint, a6 *uint) = ",
 		},
 		{
 			name:        "syscall with 15 params",
-			wantsysfunc: "syscall.Syscall15",
+			wantsysfunc: "syscall.SyscallN",
 			sig: strings.ReplaceAll(`Example(a1 *uint, a2 *uint, a3 *uint, a4 *uint, a5 *uint, a6 *uint,
 						a7 *uint, a8 *uint, a9 *uint, a10 *uint, a11 *uint, a12 *uint,
 						a13 *uint, a14 *uint, a15 *uint) = `, "\n", ""),
